major timeout/error handling changes (things might break)
it turns out that the normal infinite reboot method used by
all hacked SMCs hides legitimate RRoDs and keeps rebooting.
these changes now distinguish between when a timeout happens
(and we should keep rebooting), and when a real error case
happens.

i'm a little lazy and not in the mood to dig my test consoles out
of the bin, so any bugs that pop up should be reported...